Xpetra: Remove xpetra-sup, ETI BlockCrsMatrix and Matrix
