+++ packages/stokhos/src/CMakeLists.txt
@@ -948,7 +948,6 @@ IF (Stokhos_ENABLE_Sacado)
       GLOBAL_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES  )
       IF (Xpetra_ENABLE_EXPLICIT_INSTANTIATION)
         INCLUDE(${CMAKE_CURRENT_SOURCE_DIR}/../../xpetra/src/Utils/ExplicitInstantiation/ETI_SC_LO_GO_NO_classes.cmake)
-        INCLUDE(${CMAKE_CURRENT_SOURCE_DIR}/../../xpetra/sup/Utils/ExplicitInstantiation/ETI_SC_LO_GO_NO_classes.cmake)
         INCLUDE(${CMAKE_CURRENT_SOURCE_DIR}/../../xpetra/src/Utils/ExplicitInstantiation/ETI_MULTIVECTOR_classes.cmake)
         SET(XPETRA_ALL_ETI_CLASSES ${XPETRA_SC_LO_GO_NO_ETI_CLASSES} ${XPETRA_MULTIVECTOR_ETI_CLASSES})
       ENDIF()
