+++ packages/xpetra/CMakeLists.txt
@@ -193,7 +193,6 @@ ENDIF()
 #
 
 ADD_SUBDIRECTORY(src)
-ADD_SUBDIRECTORY(sup)
 
 IF (IS_DIRECTORY ${CMAKE_CURRENT_SOURCE_DIR}/doc
     AND EXISTS ${CMAKE_CURRENT_SOURCE_DIR}/doc/CMakeLists.txt)
