+++ packages/xpetra/src/StridedMap/Xpetra_StridedMapFactory_def.hpp
