+++ packages/xpetra/src/Utils/ClassList/SC-LO-GO-NO.classList
@@ -1,14 +1,18 @@
-#BlockedCrsMatrix
+BlockedCrsMatrix
 #BlockedMultiVector
 #BlockedVector
 #CrsMatrix
 #CrsMatrixFactory
 #CrsMatrixUtils
+CrsMatrixWrap
 #MapExtractor
 #MapExtractorFactory
-#Matrix
-#MatrixUtils
-#MatrixFactory
+Helpers
+IO
+Matrix
+MatrixUtils
+MatrixFactory
+MatrixMatrix
 #MultiVectorFactory
 #Operator
 #ReorderedBlockedCrsMatrix
@@ -16,6 +20,7 @@ TpetraBlockCrsMatrix
 TpetraCrsMatrix
 #TpetraHalfPrecisionOperator
 ##TpetraVbrMatrix
+TripleMatrixMultiply
 ##VbrMatrix
 #Vector
 VectorFactory
