+++ packages/xpetra/src/Utils/ExplicitInstantiation/ETI_LO_GO_NO_classes.cmake
@@ -2,6 +2,8 @@
 APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::BlockedMap )
 APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::Map )
 APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::MapFactory )
+APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::StridedMap )
+APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::StridedMapFactory )
 APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::TpetraCrsGraph )
 APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::TpetraExport )
 APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::TpetraImport )
