+++ packages/xpetra/src/Utils/ExplicitInstantiation/ETI_SC_LO_GO_NO_classes.cmake
@@ -1,4 +1,13 @@
 # This file is automatically generated by xpetra/src/Utils/ExplicitInstantiation/genClassLists.sh
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::BlockedCrsMatrix )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::CrsMatrixWrap )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::Helpers )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::IO )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::Matrix )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixUtils )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixFactory )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::MatrixMatrix )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::TpetraBlockCrsMatrix )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::TpetraCrsMatrix )
+APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::TripleMatrixMultiply )
 APPEND_SET(XPETRA_SC_LO_GO_NO_ETI_CLASSES Xpetra::VectorFactory )
