--- packages/xpetra/sup/Utils/ExplicitInstantiation/ETI_LO_GO_NO_classes.cmake
@@ -1,4 +1,0 @@
-APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::StridedMap )
-APPEND_SET(XPETRA_LO_GO_NO_ETI_CLASSES Xpetra::StridedMapFactory )
-
-
